IdealCoproduct

Description

Haskell implementation of the coproduct of two ideal monads. For abstract nonsense, consult N. Ghani, T. Uustalu Coproducts of ideal monads http://www.cs.ioc.ee/~tarmo/papers/fics03-tia.pdf.

Documentation

newtype Turns h t a

Values of Turns h t a can be seen as interleaved layers of the functors h and t (they take turns) with variables of the type a possible on every level (except for the outermost layer, which is always given by h).

Constructors

Turns

Fields

unTurns :: h (Either a (Turns t h a))

data IdealCoproduct m n a

The coproduct (a disjoint sum in the category of ideal monads) of ideal monads m and n. Each value consists of alternating layers of the ideals of m and n with variables of the type a.

Constructors

ICVar a

ICLeft (Turns (Ideal m) (Ideal n) a)

ICRight (Turns (Ideal n) (Ideal m) a)

sym :: IdealCoproduct m n a -> IdealCoproduct n m a

Swap the ICLeft and ICRight constructors. The function sym is an involution, that is, sym . sym = id.

newtype AmbiTurns h t a

The ideal of IdealCoproduct.

Constructors

AmbiTurns

Fields

unAmbiTurns :: Either (Turns h t a) (Turns t h a)

liftl :: (Functor (Ideal m), Functor (Ideal n), MonadIdeal m, MonadIdeal n) => m a -> IdealCoproduct m n a

Lift an ideal monad m to IdealCoproduct m n. The function inl respects the equations of the MonadTrans class, that is:

liftl . return  =  return

liftl m >>= liftl . f  =  liftl (m >>= f)

liftr :: (Functor (Ideal m), Functor (Ideal n), MonadIdeal m, MonadIdeal n) => n a -> IdealCoproduct m n a

Symmetric version of liftl.

foldTurns :: (Functor h, Functor t) => (h a -> a) -> (t a -> a) -> Turns h t a -> a

Fold the structure of a Turns using an h-algebra and a t-algebra.

foldIdealCoproduct :: (Functor (Ideal m), Functor (Ideal n), MonadIdeal m, MonadIdeal n) => (m a -> a) -> (n a -> a) -> IdealCoproduct m n a -> a

Fold the structure of an IdealCoproduct using an m-algebra and a n-algebra.

interpTurns :: Monad k => (forall a. h a -> k a) -> (forall a. t a -> k a) -> Turns h t a -> k a

Fold the structure of a Turns by interpreting each level as a computation in a monad k and then join-ing them to obtain one computation in k.

interpIdealCoproduct :: (Functor (Ideal m), Functor (Ideal n), MonadIdeal m, MonadIdeal n, Monad k) => (forall a. m a -> k a) -> (forall a. n a -> k a) -> IdealCoproduct m n a -> k a

Fold the structure of an IdealCoproduct by interpreting each level as a computation in a monad k and then join-ing them to obtain one computation in k. From the category-theoretic point of view, if the first two arguments are ideal monad morphisms, interpIdealCoproduct is the coproduct mediator.