Instances

Non-empty list

data AtLeast2 a

List with at least two elements

Constructors

a :|: (NonEmpty a)

Free monad

newtype Wrap f a

Free monad generated by a functor f wrapped in f. In other words, it is the type of "free monads" with at least one layer of f.

Constructors

Wrap

Fields

unWrap :: f (Free f a)