Resumption

Description

Generalised resumption monad a la M. Piróg, N. Wu, J. Gibbons Modules over monads and their algebras https://coalg.org/calco15/papers/p18-Piróg.pdf.

Documentation

newtype Resumption m r a

The generalised resumption monad is a composition of a monad m and the free monad generated by r (intended to be a right module over m).

Constructors

Resumption

Fields

unResumption :: m (Free r a)

newtype MWrap m r a

Type of resumptions with at least one level of free structure.

Constructors

MWrap

Fields

unMWrap :: m (Wrap r a)

liftm :: (Functor m, RModule m r) => m a -> Resumption m r a

Lifts a computation in a monad m to a computation in the resumption monad.

liftr :: (Functor m, RModule m r) => r a -> Resumption m r a

Lifts a value of a right module over a monad m to a computation in the resumption monad.

foldResumption :: (Functor m, Functor r) => (m a -> a) -> (r a -> a) -> Resumption m r a -> a

Fold the structure of a resumption using an m-algebra and and r-algebra.

interpResumption :: (Functor k, Monad k) => (forall a. m a -> k a) -> (forall a. r a -> k a) -> Resumption m r a -> k a

Fold the structure of a resumption by interpreting each layer as a computation in a monad k and then join-ing the layers.